vollständige implementierung von copy

+++ Verzeichnis.py
@@ -2,6 +2,7 @@ import logging
 import os
 import pyexiv2
 import string
+import shutil
 
 def search(path):
     logging.info('search for images in %s' % path)
@@ -41,41 +42,43 @@ def Exif(iPath):
     metadata.exif_keys
 
     # Blende anzeigen
-    tagNr = metadata['Exif.Photo.FNumber']
-    print tagNr.raw_value
+    #tagNr = metadata['Exif.Photo.FNumber']
+    #print tagNr.raw_value
     # Verschlusszeit anzeigen
-    tagExp = metadata['Exif.Photo.ExposureTime']
-    print tagExp.raw_value
+    #tagExp = metadata['Exif.Photo.ExposureTime']
+    #print tagExp.raw_value
     # ISO anzeigen
-    tagISO = metadata['Exif.Photo.ISOSpeedRatings']
-    print tagISO.raw_value
+    #tagISO = metadata['Exif.Photo.ISOSpeedRatings']
+    #print tagISO.raw_value
     # Kameramodel anzeigen
     tagMdl = metadata['Exif.Image.Model']
     print tagMdl.raw_value
     # Aufnahmedatum anzeigen
     tagTime = metadata['Exif.Image.DateTime']
-    print tagTime.value.year
-    print tagTime.raw_value
+    #print tagTime.value.year
+    print tagTime.value
 
-def copy(source, dest):
-    dest = 'C:/Users/paull/Documents/'
- #if os.path.exists(dest) == False:
-    print(dest)
-    print(source)
+
+def copy( source, dest):
+
+ cfile = os.path.basename(source)
+ destfile = os.path.join(dest, cfile)
+ if os.path.exists(destfile) == False:
     shutil.copy2(source, dest)
     sourceStat = os.stat(source)
-    destStat = os.stat(dest)
-    #if destStat.st_size == sourceStat.st_size:
-    os.remove(source)
-    print('File erfolgreich kopiert')
-    #else:
-        #print('Kopiervorgang fehlgschlagen')
- #else:
-#     print('File bereits vorhanden')
+    destStat = os.stat(destfile)
+    if destStat.st_size == sourceStat.st_size:
+        os.remove(source)
+        print('File erfolgreich kopiert')
+    else:
+        print('Kopiervorgang fehlgschlagen')
+ else:
+     print('File bereits vorhanden')
 
 
 logging.basicConfig(level=logging.DEBUG)
 logging.info("starting")
+dest = 'C:/Users/paull/Documents/'
 for dir in dirSearch():
  #path = dir
  #fl = []
@@ -83,4 +86,4 @@ for dir in dirSearch():
  for file in files:
    print file
    Exif(file)
-   copy(file, '')
+   copy(file, dest)
